requirements.txt
requests
python-dotenv
python-dateutil
ollama